(be): Break input-validation into different files
